fix: move test in 201702897 directory
